+++ lib/tib_api.dart
@@ -10,7 +10,8 @@ library;
 // APIs
 export 'src/api/departures.dart';
 export 'src/api/stations.dart';
-export 'src/api/line.dart';
+export 'src/api/station_line.dart';
+export 'src/api/route_line.dart';
 
 // Realtime
 export 'src/realtime/bus_position.dart';
@@ -19,3 +20,6 @@ export 'src/realtime/station_info.dart';
 
 // Sockets
 export 'src/sockets/location_socket.dart';
+
+// Messaging
+export 'src/messaging/tib_rss.dart';