+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 0.3.0
+
+- Added RSS Support
+- Added scraping of the TIB warnings
+- Added Realtime ID to the RealTime object
+- Updated Docs
+- Added Support for querying Lines & Line Paths
+
 ## 0.2.0
 
 - Error handling